Refactored resolvable module to a separate file
