+++ external-crates/move/crates/move-compiler/src/naming/translate.rs
@@ -18,6 +18,7 @@ use crate::{
     naming::{
         ast::{self as N, BlockLabel, NominalBlockUsage, TParamID},
         fake_natives,
+        resolvable_module::ResolvableModule,
         syntax_methods::resolve_syntax_attributes,
     },
     parser::ast::{
@@ -25,7 +26,7 @@ use crate::{
     },
     shared::{
         ide::{EllipsisMatchEntries, IDEAnnotation, IDEInfo},
-        program_info::{ModuleInfo, NamingProgramInfo},
+        program_info::NamingProgramInfo,
         unique_map::UniqueMap,
         *,
     },
@@ -203,28 +204,6 @@ enum TypeAnnotation {
     FunctionSignature,
     MacroSignature,
     Expression,
-}
-
-//**************************************************************************************************
-// Resolvable Module
-//**************************************************************************************************
-
-/// Trait to abstract over module-like structures for member resolution
-trait ResolvableModule {
-    /// Returns iterator over structs: (name, type_params_len, field_info, loc)
-    fn structs(&self) -> impl Iterator<Item = (DatatypeName, usize, FieldInfo, Loc)>;
-
-    /// Returns iterator over enums: (name, type_params_len, loc, variants)
-    /// where variants is Vec of (variant_name, field_info, variant_loc)
-    fn enums(
-        &self,
-    ) -> impl Iterator<Item = (DatatypeName, usize, Loc, Vec<(VariantName, FieldInfo, Loc)>)>;
-
-    /// Returns iterator over functions: (name, type_params_len, params_len)
-    fn functions(&self) -> impl Iterator<Item = (FunctionName, usize, usize)>;
-
-    /// Returns iterator over constants: (name, defined_loc)
-    fn constants(&self) -> impl Iterator<Item = (ConstantName, Loc)>;
 }
 
 //************************************************
@@ -415,118 +394,6 @@ impl std::fmt::Display for ResolvedModuleMember {
             ResolvedModuleMember::Constant(_) => write!(f, "constant"),
             ResolvedModuleMember::Datatype(ty) => write!(f, "{}", ty.datatype_kind_str()),
         }
-    }
-}
-
-impl ResolvableModule for E::ModuleDefinition {
-    fn structs(&self) -> impl Iterator<Item = (DatatypeName, usize, FieldInfo, Loc)> {
-        self.structs.key_cloned_iter().map(|(name, sdef)| {
-            let field_info = match &sdef.fields {
-                E::StructFields::Positional(fields) => FieldInfo::Positional(fields.len()),
-                E::StructFields::Named(f) => {
-                    FieldInfo::Named(f.key_cloned_iter().map(|(k, _)| k).collect())
-                }
-                E::StructFields::Native(_) => FieldInfo::Empty,
-            };
-            (name, sdef.type_parameters.len(), field_info, name.loc())
-        })
-    }
-
-    fn enums(
-        &self,
-    ) -> impl Iterator<Item = (DatatypeName, usize, Loc, Vec<(VariantName, FieldInfo, Loc)>)> {
-        self.enums.key_cloned_iter().map(|(enum_name, edef)| {
-            let variants: Vec<_> = edef
-                .variants
-                .key_cloned_iter()
-                .map(|(vname, vdef)| {
-                    let field_info = match &vdef.fields {
-                        E::VariantFields::Named(fields) => {
-                            FieldInfo::Named(fields.key_cloned_iter().map(|(k, _)| k).collect())
-                        }
-                        E::VariantFields::Positional(tys) => FieldInfo::Positional(tys.len()),
-                        E::VariantFields::Empty => FieldInfo::Empty,
-                    };
-                    (vname, field_info, vdef.loc)
-                })
-                .collect();
-            (enum_name, edef.type_parameters.len(), edef.loc, variants)
-        })
-    }
-
-    fn functions(&self) -> impl Iterator<Item = (FunctionName, usize, usize)> {
-        self.functions.key_cloned_iter().map(|(name, fun)| {
-            (
-                name,
-                fun.signature.type_parameters.len(),
-                fun.signature.parameters.len(),
-            )
-        })
-    }
-
-    fn constants(&self) -> impl Iterator<Item = (ConstantName, Loc)> {
-        self.constants
-            .key_cloned_iter()
-            .map(|(name, _)| (name, name.loc()))
-    }
-}
-
-impl ResolvableModule for ModuleInfo {
-    fn structs(&self) -> impl Iterator<Item = (DatatypeName, usize, FieldInfo, Loc)> {
-        self.structs.key_cloned_iter().map(|(name, sdef)| {
-            let field_info = match &sdef.fields {
-                N::StructFields::Defined(positional, fields) => {
-                    if *positional {
-                        FieldInfo::Positional(fields.len())
-                    } else {
-                        FieldInfo::Named(fields.key_cloned_iter().map(|(k, _)| k).collect())
-                    }
-                }
-                N::StructFields::Native(_) => FieldInfo::Empty,
-            };
-            (name, sdef.type_parameters.len(), field_info, name.loc())
-        })
-    }
-
-    fn enums(
-        &self,
-    ) -> impl Iterator<Item = (DatatypeName, usize, Loc, Vec<(VariantName, FieldInfo, Loc)>)> {
-        self.enums.key_cloned_iter().map(|(enum_name, edef)| {
-            let variants: Vec<_> = edef
-                .variants
-                .key_cloned_iter()
-                .map(|(vname, vdef)| {
-                    let field_info = match &vdef.fields {
-                        N::VariantFields::Defined(positional, fields) => {
-                            if *positional {
-                                FieldInfo::Positional(fields.len())
-                            } else {
-                                FieldInfo::Named(fields.key_cloned_iter().map(|(k, _)| k).collect())
-                            }
-                        }
-                        N::VariantFields::Empty => FieldInfo::Empty,
-                    };
-                    (vname, field_info, vdef.loc)
-                })
-                .collect();
-            (enum_name, edef.type_parameters.len(), edef.loc, variants)
-        })
-    }
-
-    fn functions(&self) -> impl Iterator<Item = (FunctionName, usize, usize)> {
-        self.functions.key_cloned_iter().map(|(name, finfo)| {
-            (
-                name,
-                finfo.signature.type_parameters.len(),
-                finfo.signature.parameters.len(),
-            )
-        })
-    }
-
-    fn constants(&self) -> impl Iterator<Item = (ConstantName, Loc)> {
-        self.constants
-            .key_cloned_iter()
-            .map(|(name, cinfo)| (name, cinfo.defined_loc))
     }
 }
